mqtt package got an abtstract integration test class from which other test classes can extend
