+++ communication/mqtt/src/test/java/org/openbase/jul/communication/mqtt/SharedMqttClientTest.kt
@@ -4,11 +4,11 @@ import org.junit.jupiter.api.Test
 
 import org.openbase.jul.communication.config.CommunicatorConfig
 
-internal class SharedMqttClientTest {
+internal class SharedMqttClientTest : AbstractIntegrationTest() {
 
     @Test
     fun `shutdown should be work as expected`() {
-        val config = CommunicatorConfig(IntegrationTest.broker.host, IntegrationTest.broker.firstMappedPort)
+        val config = CommunicatorConfig(brokerHost, brokerPort)
         val client = SharedMqttClient
         client.get(config).apply {
             disconnect()
